move global config to test_helper
