+++ test/test_helper.rb
@@ -3,7 +3,6 @@
 ENV['RAILS_ENV'] ||= 'test'
 require_relative '../config/environment'
 require 'rails/test_help'
-require_relative './helpers/omni_auth_helper.rb'
 
 class ActiveSupport::TestCase
   # Run tests in parallel with specified workers
@@ -18,7 +17,19 @@ end
 
 class ActionDispatch::IntegrationTest
   include Devise::Test::IntegrationHelpers
-  include OmniAuthHelper
+
+  OmniAuth.config.test_mode = true
+  OmniAuth.config.add_mock(
+    :github,
+    provider: 'github',
+    uid: '12345',
+    info: { name: 'Github User', email: 'github@github.com' }
+  )
+
+  def sign_with_github
+    Rails.application.env_config['omniauth.auth'] = OmniAuth.config.mock_auth[:github]
+    get user_github_omniauth_callback_path
+  end
 
   # def login(user)
   #   params = {
